cleanup(rewrite) gathering CQL constants in cassandra.types
