+++ spec/unit/cql_types_buffer_spec.lua
@@ -1,3 +1,4 @@
+local cassandra = require "cassandra"
 local Buffer = require "cassandra.buffer"
 local CONSTS = require "cassandra.constants"
 local types = require "cassandra.types"
@@ -36,7 +37,7 @@ describe("CQL Types protocol v"..protocol_version, function()
     describe("manual type infering", function()
       it("should be possible to infer the type of a value through helper methods", function()
         for _, fixture in ipairs(fixture_values) do
-          local infered_value = types[fixture_type](fixture)
+          local infered_value = cassandra.types[fixture_type](fixture)
           local buf = Buffer(protocol_version)
           buf:write_cql_value(infered_value)
           buf:reset()
