+++ spec/unit/requests_spec.lua
@@ -1,8 +1,7 @@
-local requests = require "cassandra.requests"
+local types = require "cassandra.types"
 local Buffer = require "cassandra.buffer"
+local requests = require "cassandra.requests"
 local frame_header = require "cassandra.types.frame_header"
-
-local op_codes = frame_header.op_codes
 
 describe("Requests", function()
   describe("StartupRequest", function()
@@ -15,7 +14,7 @@ describe("Requests", function()
       assert.equal(0x03, full_buffer:read_byte())
       assert.equal(0, full_buffer:read_byte())
       assert.equal(0, full_buffer:read_short())
-      assert.equal(op_codes.STARTUP, full_buffer:read_byte())
+      assert.equal(types.OP_CODES.STARTUP, full_buffer:read_byte())
       assert.equal(22, full_buffer:read_int())
       assert.same({CQL_VERSION = "3.0.0"}, full_buffer:read_string_map())
     end)
@@ -30,7 +29,7 @@ describe("Requests", function()
       assert.equal(0x02, full_buffer:read_byte())
       assert.equal(0, full_buffer:read_byte())
       assert.equal(0, full_buffer:read_byte())
-      assert.equal(op_codes.STARTUP, full_buffer:read_byte())
+      assert.equal(types.OP_CODES.STARTUP, full_buffer:read_byte())
       assert.equal(22, full_buffer:read_int())
       assert.same({CQL_VERSION = "3.0.0"}, full_buffer:read_string_map())
     end)
