+++ src/cassandra/types/frame_header.lua
@@ -1,6 +1,6 @@
-local utils = require "cassandra.utils.table"
 local bit = require "cassandra.utils.bit"
 local Buffer = require "cassandra.buffer"
+local table_utils = require "cassandra.utils.table"
 
 --- CONST
 -- @section constants
@@ -16,35 +16,7 @@ local VERSION_CODES = {
   }
 }
 
-setmetatable(VERSION_CODES, utils.const_mt)
-
-local FLAGS = {
-  COMPRESSION = 0x01, -- not implemented
-  TRACING = 0x02
-}
-
--- when we'll support protocol v4, other
--- flags will be added.
--- setmetatable(FLAGS, utils.const_mt)
-
-local OP_CODES = {
-  ERROR = 0x00,
-  STARTUP = 0x01,
-  READY = 0x02,
-  AUTHENTICATE = 0x03,
-  OPTIONS = 0x05,
-  SUPPORTED = 0x06,
-  QUERY = 0x07,
-  RESULT = 0x08,
-  PREPARE = 0x09,
-  EXECUTE = 0x0A,
-  REGISTER = 0x0B,
-  EVENT = 0x0C,
-  BATCH = 0x0D,
-  AUTH_CHALLENGE = 0x0E,
-  AUTH_RESPONSE = 0x0F,
-  AUTH_SUCCESS = 0x10
-}
+setmetatable(VERSION_CODES, table_utils.const_mt)
 
 --- FrameHeader
 -- @section FrameHeader
@@ -54,7 +26,7 @@ local FrameHeader = Buffer:extend()
 function FrameHeader:new(version, flags, op_code, body_length)
   self.flags = flags and flags or 0
   self.op_code = op_code
-  self.stream_id = 0 -- @TODO support streaming
+  self.stream_id = 0
   self.body_length = body_length
 
   self.super.new(self, version)
@@ -62,7 +34,7 @@ end
 
 function FrameHeader:dump()
   FrameHeader.super.write_byte(self, VERSION_CODES:get("REQUEST", self.version))
-  FrameHeader.super.write_byte(self, self.flags) -- @TODO find a more secure way
+  FrameHeader.super.write_byte(self, self.flags)
 
   if self.version < 3 then
     FrameHeader.super.write_byte(self, self.stream_id)
@@ -70,7 +42,7 @@ function FrameHeader:dump()
     FrameHeader.super.write_short(self, self.stream_id)
   end
 
-  FrameHeader.super.write_byte(self, self.op_code) -- @TODO find a more secure way
+  FrameHeader.super.write_byte(self, self.op_code)
   FrameHeader.super.write_int(self, self.body_length)
 
   return FrameHeader.super.dump(self)
@@ -107,8 +79,4 @@ function FrameHeader.from_raw_bytes(version_byte, raw_bytes)
   return FrameHeader(version, flags, op_code, body_length)
 end
 
-return {
-  op_codes = OP_CODES,
-  flags = FLAGS,
-  FrameHeader = FrameHeader
-}
+return FrameHeader
